+++ R/misc.R
@@ -1,4 +1,11 @@
 
+get_hostname <- function() {
+  hn <- try(system("hostname", intern = TRUE, ignore.stderr = TRUE), silent = TRUE)
+  if(inherits(hn, "try-error"))
+    hn <- ""
+  hn
+}
+
 get_output_index <- function() {
   server_dir <- get_server_dir()
   idx_path <- file.path(get_server_dir(), ".idx")
@@ -13,6 +20,12 @@ get_output_file <- function(ii) {
 
 set_index_template <- function() {
   a <- suppressWarnings(readLines(system.file("index.html", package = "rmote")))
+  if(getOption("rmote_hostname")) {
+    idx <- grepl("rmote viewer", a[1:10])
+    hn <- get_hostname()
+    if(length(idx) > 0 && hn != "")
+      a[idx] <- gsub("rmote viewer", paste("rmote:", hn), a[idx])
+  }
   options(rmote_index = paste(a, collapse = "\n"))
 }