+++ R/rmote.R
@@ -12,13 +12,16 @@ NULL
 #' @param graphics (logical) send traditional lattice / ggplot2 plots to servr
 #' @param basegraphics (logical) send base graphics to servr
 #' @param htmlwidgets (logical) send htmlwidgets to servr
+#' @param hostname (logical) try to get hostname and use it in viewer page title
 #'
 #' @export
 #' @importFrom servr httd
-start_rmote <- function(server_dir = file.path(tempdir(), "rmote_server"),
-  port = 4321,
-  daemon = TRUE, help = TRUE, graphics = TRUE,
-  basegraphics = TRUE, htmlwidgets = TRUE) {
+start_rmote <- function(
+  server_dir = file.path(tempdir(), "rmote_server"),
+  port = 4321, daemon = TRUE,
+  help = TRUE, graphics = TRUE,
+  basegraphics = TRUE, htmlwidgets = TRUE,
+  hostname = TRUE) {
 
   if(!file.exists(server_dir))
     dir.create(server_dir, recursive = TRUE, showWarnings = FALSE)
@@ -29,6 +32,7 @@ start_rmote <- function(server_dir = file.path(tempdir(), "rmote_server"),
   options(rmote_graphics = graphics)
   options(rmote_htmlwidgets = htmlwidgets)
   options(rmote_basegraphics = basegraphics)
+  options(rmote_hostname = hostname)
 
   set_index_template()
 
@@ -56,11 +60,13 @@ stop_rmote <- function() {
 #' @param graphics (logical) send traditional lattice / ggplot2 plots to servr
 #' @param basegraphics (logical) send base graphics to servr
 #' @param htmlwidgets (logical) send htmlwidgets to servr
+#' @param hostname (logical) try to get hostname and use it in viewer page title
 #' @note This is useful when running multiple R sessions on a server, where all will serve the same rmote process.  It is not necessary to call this in the same session on which \code{\link{start_rmote}} has been called, but on any other R sessions.
 #' @export
 rmote_on <- function(server_dir,
   help = TRUE, graphics = TRUE,
-  basegraphics = TRUE, htmlwidgets = TRUE
+  basegraphics = TRUE, htmlwidgets = TRUE,
+  hostname = TRUE
 ) {
 
   if(!file.exists(server_dir))
@@ -73,6 +79,7 @@ rmote_on <- function(server_dir,
   options(rmote_graphics = graphics)
   options(rmote_htmlwidgets = htmlwidgets)
   options(rmote_basegraphics = basegraphics)
+  options(rmote_hostname = hostname)
 
   if(basegraphics)
     set_base_plot_hook()
@@ -92,6 +99,16 @@ rmote_off <- function() {
 
 is_rmote_on <- function() {
   getOption("rmote_on", FALSE) || length(servr::daemon_list()) > 0
+}
+
+no_other_devices <- function() {
+  res <- length(dev.list()) == 0
+  if(!res) {
+    message("- not sending to rmote because another graphics device has been opened...")
+    message("- sending to the open graphics device instead...")
+    message("- to send to rmote, close all active graphics devices using graphics.off()")
+  }
+  res
 }
 
 get_server_dir <- function() {